+++ userland/include/tree.h
@@ -11,9 +11,9 @@
  */
 typedef struct {
     mac_addr_t addr;
-    uint32_t high_pwr; // the power at which the parent does currently broadcast data frames
-    uint32_t snd_high_pwr; // the power at which the parent WOULD broadacst data frames, if its furthest child were to disconnect
-    uint32_t own_pwr; // the minimum power with which the parent has to broadcast to reach us
+    int8_t high_pwr; // the power at which the parent does currently broadcast data frames
+    int8_t snd_high_pwr; // the power at which the parent WOULD broadacst data frames, if its furthest child were to disconnect
+    int8_t own_pwr; // the minimum power with which the parent has to broadcast to reach us
 } parent_t;
 
 /**
@@ -21,7 +21,7 @@ typedef struct {
  */
 typedef struct {
     mac_addr_t addr;
-    uint32_t tx_pwr; // minimum power with which we have to broadcast to reach this child
+    int8_t tx_pwr; // minimum power with which we have to broadcast to reach this child
 } child_t;
 
 /**
@@ -30,9 +30,9 @@ typedef struct {
 typedef struct {
     bool is_source; // whether we are the rood of the broadcast tree
     map_t children; // hashmap of currently connected children
-    uint32_t max_pwr; // maximum power at which we are able (or willing) to broadcast
-    uint32_t high_pwr; // the power at which we currently broadcast data frames
-    uint32_t snd_high_pwr; // the power at which we WOULD broadacst data frames, if our furthest child were to disconnect
+    int8_t max_pwr; // maximum power at which we are able (or willing) to broadcast
+    int8_t high_pwr; // the power at which we currently broadcast data frames
+    int8_t snd_high_pwr; // the power at which we WOULD broadacst data frames, if our furthest child were to disconnect
     mac_addr_t laddr; // local mac address
     parent_t *parent; // currently connected parent
     parent_t *pending_parent;
